Split out PassthroughPlatformConfiguration
